requirements.txt
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
drf-spectacular==0.13.2
Markdown==3.3.3
mysqlclient==2.0.3
